Create AuthPage component for handling authentication and redirecting after authentication.
Allows for in-place authentication. withAuth wraps the page component on export.

when the page loads we save it's location to sessionStorage using the key 'redirectAfterAuth'

and then immediately redirect to a new page /auth.

when the user puts in the appropriate password they are redirected to the appropriate page, maintaining the flow of page-transition animations.

once a user has successfully authenticated, they will not be asked again unless the session is reset (tab / window closed)